@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
 	<PropertyGroup>
-		<TargetFramework>net7.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<Nullable>enable</Nullable>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<UserSecretsId>bfd227b7-eb90-46b4-84c0-9c3a76c7cbe9</UserSecretsId>
@@ -23,15 +23,15 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Grpc.Tools" Version="2.58.0">
+		<PackageReference Include="Grpc.Tools" Version="2.59.0">
 			<PrivateAssets>all</PrivateAssets>
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
-		<PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.17.2" />
-		<PackageReference Include="MMLib.SwaggerForOcelot" Version="7.0.0" />
-		<PackageReference Include="Ocelot" Version="19.0.4" />
-		<PackageReference Include="Ocelot.Cache.CacheManager" Version="19.0.2" />
-		<PackageReference Include="Ocelot.Provider.Polly" Version="19.0.4" />
+		<PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.19.5" />
+		<PackageReference Include="MMLib.SwaggerForOcelot" Version="8.1.0" />
+		<PackageReference Include="Ocelot" Version="22.0.1" />
+		<PackageReference Include="Ocelot.Cache.CacheManager" Version="22.0.1" />
+		<PackageReference Include="Ocelot.Provider.Polly" Version="22.0.1" />
 	</ItemGroup>
 
 	<ItemGroup>
