@@ -1,83 +1,83 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net7.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Autofac" Version="7.0.1" />
+		<PackageReference Include="Autofac" Version="7.1.0" />
 		<PackageReference Include="Autofac.Extensions.DependencyInjection" Version="8.0.0" />
-		<PackageReference Include="AutoFixture" Version="4.18.0" />
-		<PackageReference Include="Consul" Version="1.6.10.9" />
-		<PackageReference Include="Dapper" Version="2.0.143" />
-		<PackageReference Include="DotNetCore.CAP" Version="7.2.1" />
-		<PackageReference Include="DotNetCore.CAP.Dashboard" Version="7.2.1" />
-		<PackageReference Include="DotNetCore.CAP.RabbitMQ" Version="7.2.1" />
-		<PackageReference Include="DotNetCore.CAP.SqlServer" Version="7.2.1" />
+		<PackageReference Include="AutoFixture" Version="4.18.1" />
+		<PackageReference Include="Consul" Version="1.7.14.2" />
+		<PackageReference Include="Dapper" Version="2.1.24" />
+		<PackageReference Include="DotNetCore.CAP" Version="7.2.2" />
+		<PackageReference Include="DotNetCore.CAP.Dashboard" Version="7.2.2" />
+		<PackageReference Include="DotNetCore.CAP.RabbitMQ" Version="7.2.2" />
+		<PackageReference Include="DotNetCore.CAP.SqlServer" Version="7.2.2" />
 		<PackageReference Include="EventBus.Base.Standard" Version="1.0.2" />
 		<PackageReference Include="EventBus.RabbitMQ.Standard" Version="3.1.0" />
 		<PackageReference Include="Grpc.AspNetCore" Version="2.37.0" />
-		<PackageReference Include="Grpc.AspNetCore.Web" Version="2.55.0" />
+		<PackageReference Include="Grpc.AspNetCore.Web" Version="2.59.0" />
 		<PackageReference Include="Grpc.Core" Version="2.46.6" />
 		<PackageReference Include="Grpc.Core.Testing" Version="2.46.6" />
 		<PackageReference Include="Grpc.Net.Client" Version="2.37.0" />
-		<PackageReference Include="Grpc.Net.Client.Web" Version="2.55.0" />
-		<PackageReference Include="Grpc.Tools" Version="2.56.0">
+		<PackageReference Include="Grpc.Net.Client.Web" Version="2.59.0" />
+		<PackageReference Include="Grpc.Tools" Version="2.59.0">
 			<PrivateAssets>all</PrivateAssets>
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
-		<PackageReference Include="Hangfire" Version="1.8.3" />
-		<PackageReference Include="Hangfire.AspNetCore" Version="1.8.3" />
-		<PackageReference Include="Hangfire.Core" Version="1.8.3" />
+		<PackageReference Include="Hangfire" Version="1.8.6" />
+		<PackageReference Include="Hangfire.AspNetCore" Version="1.8.6" />
+		<PackageReference Include="Hangfire.Core" Version="1.8.6" />
 		<PackageReference Include="IdGen.DependencyInjection" Version="3.0.0" />
 		<PackageReference Include="Jaeger" Version="1.0.3" />
-		<PackageReference Include="MediatR" Version="12.1.1" />
+		<PackageReference Include="MediatR" Version="12.2.0" />
 		<PackageReference Include="MediatR.Contracts" Version="2.0.1" />
-		<PackageReference Include="MediatR.Extensions.Autofac.DependencyInjection" Version="11.0.0" />
-		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="7.0.9" />
+		<PackageReference Include="MediatR.Extensions.Autofac.DependencyInjection" Version="11.3.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="8.0.0" />
 		<PackageReference Include="Microsoft.AspNetCore.Diagnostics" Version="2.2.0" />
 		<PackageReference Include="Microsoft.AspNetCore.HttpsPolicy" Version="2.2.0" />
 		<PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.2.5" />
-		<PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="7.0.9" />
+		<PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="8.0.0" />
 		<PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning" Version="5.1.0" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.9" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="7.0.9" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="7.0.9" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.9" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="7.0.9">
+		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.0" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="8.0.0" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="8.0.0" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="8.0.0" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="8.0.0">
 			<PrivateAssets>all</PrivateAssets>
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="7.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Options" Version="7.0.1" />
-		<PackageReference Include="Microsoft.Extensions.Caching.StackExchangeRedis" Version="7.0.9" />
-		<PackageReference Include="Microsoft.Extensions.Identity.Core" Version="7.0.9" />
-		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.6.3" />
-		<PackageReference Include="Moq" Version="4.18.4" />
-		<PackageReference Include="NLog.Web.AspNetCore" Version="5.3.2" />
-		<PackageReference Include="Ocelot.Provider.Consul" Version="19.0.4" />
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Logging" Version="8.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Options" Version="8.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Caching.StackExchangeRedis" Version="8.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Identity.Core" Version="8.0.0" />
+		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0" />
+		<PackageReference Include="Moq" Version="4.20.70" />
+		<PackageReference Include="NLog.Web.AspNetCore" Version="5.3.7" />
+		<PackageReference Include="Ocelot.Provider.Consul" Version="22.0.1" />
 		<PackageReference Include="OpenTracing" Version="0.12.1" />
 		<PackageReference Include="OpenTracing.Contrib.NetCore" Version="0.9.0" />
 		<PackageReference Include="Pluralize.NET" Version="1.0.2" />
-		<PackageReference Include="Polly" Version="7.2.4" />
-		<PackageReference Include="RabbitMQ.Client" Version="6.5.0" />
-		<PackageReference Include="StackExchange.Redis" Version="2.6.122" />
+		<PackageReference Include="Polly" Version="8.2.0" />
+		<PackageReference Include="RabbitMQ.Client" Version="6.8.0" />
+		<PackageReference Include="StackExchange.Redis" Version="2.7.4" />
 		<PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
 		<PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="6.5.0" />
-		<PackageReference Include="Swashbuckle.AspNetCore.Filters" Version="7.0.6" />
-		<PackageReference Include="Grpc.AspNetCore" Version="2.55.0" />
-		<PackageReference Include="Google.Protobuf" Version="3.23.4" />
-		<PackageReference Include="Grpc.Net.Client" Version="2.55.0" />
+		<PackageReference Include="Swashbuckle.AspNetCore.Filters" Version="7.0.12" />
+		<PackageReference Include="Grpc.AspNetCore" Version="2.59.0" />
+		<PackageReference Include="Google.Protobuf" Version="3.25.1" />
+		<PackageReference Include="Grpc.Net.Client" Version="2.59.0" />
 		<PackageReference Include="Qoollo.Turbo" Version="3.1.0" />
-		<PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="11.1.0" />
-		<PackageReference Include="Z.EntityFramework.Plus.EFCore" Version="7.22.4" />
+		<PackageReference Include="Z.EntityFramework.Plus.EFCore" Version="8.101.1.1" />
 	</ItemGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\ErSoftDev.Common\ErSoftDev.Common.csproj" />
-		<ProjectReference Include="..\ErSoftDev.DomainSeedWork\ErSoftDev.DomainSeedWork.csproj" />
+	  <ProjectReference Include="..\ErSoftDev.Common\ErSoftDev.Common.csproj" />
+	  <ProjectReference Include="..\ErSoftDev.DomainSeedWork\ErSoftDev.DomainSeedWork.csproj" />
 	</ItemGroup>
 
 </Project>
