@@ -1,19 +1,19 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net8.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <Folder Include="DomainEventHandlers\" />
-    <Folder Include="IntegrationEvents\" />
-  </ItemGroup>
+	<ItemGroup>
+		<Folder Include="DomainEventHandlers\" />
+		<Folder Include="IntegrationEvents\" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\ErSoftDev.Identity.Domain\ErSoftDev.Identity.Domain.csproj" />
-    <ProjectReference Include="..\ErSoftDev.Identity.Infrastructure\ErSoftDev.Identity.Infrastructure.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\ErSoftDev.Identity.Domain\ErSoftDev.Identity.Domain.csproj" />
+		<ProjectReference Include="..\ErSoftDev.Identity.Infrastructure\ErSoftDev.Identity.Infrastructure.csproj" />
+	</ItemGroup>
 
 </Project>
