@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net7.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
 	</PropertyGroup>
@@ -11,7 +11,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <ProjectReference Include="..\..\ErSoftDev.DomainSeedWork\ErSoftDev.DomainSeedWork.csproj" />
+		<ProjectReference Include="..\..\ErSoftDev.DomainSeedWork\ErSoftDev.DomainSeedWork.csproj" />
 	</ItemGroup>
 
 </Project>
