@@ -1,14 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net8.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\..\ErSoftDev.Framework\ErSoftDev.Framework.csproj" />
-    <ProjectReference Include="..\ErSoftDev.Identity.Domain\ErSoftDev.Identity.Domain.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\..\ErSoftDev.Framework\ErSoftDev.Framework.csproj" />
+		<ProjectReference Include="..\ErSoftDev.Identity.Domain\ErSoftDev.Identity.Domain.csproj" />
+	</ItemGroup>
 
 </Project>
