@@ -1,4 +1,5 @@
 ﻿using System.Reflection;
+using ErSoftDev.Common.Utilities;
 using ErSoftDev.Framework.Grpc;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Routing;
@@ -9,8 +10,12 @@ namespace ErSoftDev.Framework.Configuration
     {
         public static void UseGrpcEndPoint(this IEndpointRouteBuilder routeBuilder)
         {
-            var grpcServices = AppDomain.CurrentDomain
-                .GetAssemblies()
+
+            var assemblies = Tools.GetAllAssemblies();
+            var grpcServices =
+                //AppDomain.CurrentDomain
+                //.GetAssemblies()
+                assemblies
                 .SelectMany(a => a.DefinedTypes)
                 .Where(t => typeof(IGrpcService).IsAssignableFrom(t) && !t.IsInterface && t.IsClass && t.IsPublic);
 
